rename: IndexPeopleToMeilisearchCommand
- also formatting and docblocks
